Refactor(exporter): decouple Decided v1 from HTTP framework
